+++ MandelbrotViewer/AppVersion.cs
@@ -8,8 +8,8 @@ namespace MandelbrotViewer;
 public static class AppVersion
 {
     public const int Major = 2;
-    public const int Minor = 0;
-    public const int Patch = 1;
+    public const int Minor = 1;
+    public const int Patch = 0;
 
     public static string Full => $"{Major}.{Minor}.{Patch}";